chore(tests): improve structure
