+++ python/pytest.ini
@@ -1,0 +1,5 @@
+[pytest]
+testpaths = tests/e2e
+python_files = test_*.py
+python_functions = test_*
+addopts = -v --strict-markers
