+++ python/tests/conftest.py
@@ -1,0 +1,4 @@
+pytest_plugins = [
+    "e2e.fixtures.client",
+    "e2e.fixtures.server",
+]
